refactor configuration of tyre info overlay
